new update
This update was supposed to be in the place of the last one posted, but I uploaded an old version by accident (oops). The "RES SCOREBOARD" now functions properly (without restart), but if the top is filled by scores of 0000, the last ones will take the name of the first or second place, not a big problem, but it is not perfect.